(Really!!) Completed RtpDump Reader and Writer (All Formats Except Short form 100%) Reverse Engineered bark.rtp and made my files match the output. The spec on the site is out of date and the origional author of the tools does not have time to look into why this is.
Added RtpDump Modificiation Test.

Added RtpDump Header Only Test.

Added RtpDump File Comparison Test.

Updated Utility-
Made GetBytesUnsafe for effiency.
Removed TicksPerMicroSecond constant.

Updated RtpPacket to allow only header and changed size of all fields to byte rather than int. Changed SequenceNumber to ushort.

Updated other classes to support changes as stated.

+++ Rtsp/Server/Streams/ImageStreams.cs
@@ -22,7 +22,7 @@ namespace Media.Rtsp.Server.Streams
         protected uint timeStamp = 0;
 
         //Should be moved to SourceStream?
-        protected uint sequenceNumber = 0;
+        protected ushort sequenceNumber = 0;
 
         //Where images are placed to be encoded and sent (should have a Dictionary<string, Queue<Image>> if more then one track should be supported.)
         //Or MediaDescription, Queue<Image>
@@ -269,7 +269,7 @@ namespace Media.Rtsp.Server.Streams
                     {
                         packet.Channel = interleave.DataChannel;
                         packet.TimeStamp = timeStamp;
-                        packet.SequenceNumber = (int)++sequenceNumber;
+                        packet.SequenceNumber = ++sequenceNumber;
                         next.Add(packet);
 
                         //Fire an event so the server sends a packet to all clients connected to this source
